change the slp_global_lock to the tgl
We do not use slp_global_lock device anymore.
We use tgl device instead and
the use space interface is changed because of this change.

When the kernel version is up and kernel can use
the dma_fence, libtbm-exynos does not use the tgl for
the bo lock and for the bo cache_flush anymore.

Change-Id: I738338b0834f6d84b94d84552509f7d9073f4429